core: replace old files with new api code
