--- Andalusian.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32407.337
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Andalusian", "Andalusian\Andalusian.csproj", "{C8495266-1898-4469-A8F1-ECCF43EC8EE7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C8495266-1898-4469-A8F1-ECCF43EC8EE7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C8495266-1898-4469-A8F1-ECCF43EC8EE7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C8495266-1898-4469-A8F1-ECCF43EC8EE7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C8495266-1898-4469-A8F1-ECCF43EC8EE7}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2D7B783D-C0A5-471F-A00F-1EABFACB3BCA}
-	EndGlobalSection
-EndGlobal
