--- Andalusian/Main.cs
@@ -1,17 +1,0 @@
-﻿using BepInEx;
-
-namespace Andalusian
-{
-    [BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
-    [BepInDependency("com.damocles.blasphemous.modding-api", "1.4.0")]
-    [BepInProcess("Blasphemous.exe")]
-    public class Main : BaseUnityPlugin
-    {
-        public static Andalusia Andalusia { get; private set; }
-
-        private void Start()
-        {
-            Andalusia = new Andalusia(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION);
-        }
-    }
-}
