+++ Blasphemous.Andalusian.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34408.163
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blasphemous.Andalusian", "Blasphemous.Andalusian\Blasphemous.Andalusian.csproj", "{74CB261A-383F-427F-9D64-186CBCC03013}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{74CB261A-383F-427F-9D64-186CBCC03013}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{74CB261A-383F-427F-9D64-186CBCC03013}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{74CB261A-383F-427F-9D64-186CBCC03013}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{74CB261A-383F-427F-9D64-186CBCC03013}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4B205DD9-7354-4D09-8FDF-1521533F1A84}
+	EndGlobalSection
+EndGlobal
